Refactor: Refactor mobileFlag
